backend: store video timestamps as decimal

+++ backend/endpoints_rooms.go
@@ -164,7 +164,7 @@ type PlayerStateMessageBi struct {
 type PlayerStateMessageData struct {
 	Paused     bool      `json:"paused"`
 	Speed      int       `json:"speed"`
-	Timestamp  int       `json:"timestamp"`
+	Timestamp  float64   `json:"timestamp"`
 	LastAction time.Time `json:"lastAction"`
 }
